+++ bbmod/src/imgui_dinput8_dev.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#define DIRECTINPUT_VERSION 0x0800
 #include <dinput.h>
 
 class ImguiDInput : IDirectInput8A {