+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Custom imgui theme support (@Solybum #18). Theme is loaded from `addons/theme.ini`
 - `get_cwd` returns a string of the cwd (@jtuu #17)
 - `play_sound` invokes winmm PlaySound with given path, ignoring errors (@jtuu #17)
 - `is_pso_focused` returns a boolean of whether or not the PSOBB HWND
   is the active window. (@jtuu #17)
+
+
 
 ## v0.3.4